@@ -13,9 +13,9 @@ dependencies:
   flutter:
     sdk: flutter
   responsive_builder: ^0.7.0
-  stream_chat_flutter: ^6.4.0
-  stream_chat_localizations: ^5.4.0
-  stream_chat_persistence: ^6.4.0
+  stream_chat_flutter: ^7.2.0
+  stream_chat_localizations: ^7.2.0
+  stream_chat_persistence: ^7.2.0
 
 flutter:
   uses-material-design: true
