@@ -12,8 +12,8 @@ dependencies:
   cupertino_icons: ^1.0.3
   flutter:
     sdk: flutter
-  stream_chat_flutter: ^6.4.0
-  stream_chat_localizations: ^5.4.0
+  stream_chat_flutter: ^7.2.0
+  stream_chat_localizations: ^7.2.0
 
 flutter:
   uses-material-design: true
