@@ -321,7 +321,7 @@ Future<T?> showChannelInfoModalBottomSheet<T>({
 ///    sheet.
 ///  * [Scaffold.of], for information about how to obtain the [BuildContext].
 ///  * <https://material.io/design/components/sheets-bottom.html#standard-bottom-sheet>
-PersistentBottomSheetController<T> showChannelInfoBottomSheet<T>({
+PersistentBottomSheetController showChannelInfoBottomSheet<T>({
   required BuildContext context,
   required Channel channel,
   Color? backgroundColor,
@@ -336,7 +336,7 @@ PersistentBottomSheetController<T> showChannelInfoBottomSheet<T>({
   VoidCallback? onDeleteConversationTap,
   VoidCallback? onCancelTap,
 }) =>
-    showBottomSheet<T>(
+    showBottomSheet(
       context: context,
       backgroundColor: backgroundColor,
       elevation: elevation,
