@@ -39,6 +39,7 @@ class StreamChat extends StatefulWidget {
     this.onBackgroundEventReceived,
     this.backgroundKeepAlive = const Duration(minutes: 1),
     this.connectivityStream,
+    this.useMaterial3 = false,
   });
 
   /// Client to do chat operations with
@@ -66,6 +67,10 @@ class StreamChat extends StatefulWidget {
   /// Visible for testing
   @visibleForTesting
   final Stream<ConnectivityResult>? connectivityStream;
+
+  /// Whether to use material 3 or not (default is false)
+  /// See our [docs](https://getstream.io/chat/docs/sdk/flutter/stream_chat_flutter/stream_chat_and_theming)
+  final bool useMaterial3;
 
   @override
   StreamChatState createState() => StreamChatState();
@@ -107,40 +112,45 @@ class StreamChatState extends State<StreamChat> {
   @override
   Widget build(BuildContext context) {
     final theme = _getTheme(context, widget.streamChatThemeData);
-    return Portal(
-      child: StreamChatConfiguration(
-        data: streamChatConfigData,
-        child: StreamChatTheme(
-          data: theme,
-          child: Builder(
-            builder: (context) {
-              final materialTheme = Theme.of(context);
-              final streamTheme = StreamChatTheme.of(context);
-              return Theme(
-                data: materialTheme.copyWith(
-                  primaryIconTheme: streamTheme.primaryIconTheme,
-                  colorScheme: materialTheme.colorScheme.copyWith(
-                    secondary: streamTheme.colorTheme.accentPrimary,
+    return Theme(
+      data: Theme.of(context).copyWith(
+        useMaterial3: widget.useMaterial3,
+      ),
+      child: Portal(
+        child: StreamChatConfiguration(
+          data: streamChatConfigData,
+          child: StreamChatTheme(
+            data: theme,
+            child: Builder(
+              builder: (context) {
+                final materialTheme = Theme.of(context);
+                final streamTheme = StreamChatTheme.of(context);
+                return Theme(
+                  data: materialTheme.copyWith(
+                    primaryIconTheme: streamTheme.primaryIconTheme,
+                    colorScheme: materialTheme.colorScheme.copyWith(
+                      secondary: streamTheme.colorTheme.accentPrimary,
+                    ),
                   ),
-                ),
-                child: StreamChatCore(
-                  client: client,
-                  onBackgroundEventReceived: widget.onBackgroundEventReceived,
-                  backgroundKeepAlive: widget.backgroundKeepAlive,
-                  connectivityStream: widget.connectivityStream,
-                  child: Builder(
-                    builder: (context) {
-                      StreamChatClient.additionalHeaders = {
-                        'X-Stream-Client':
-                            '${StreamChatClient.defaultUserAgent}-'
-                                'ui-${StreamChatClient.packageVersion}',
-                      };
-                      return widget.child ?? const Offstage();
-                    },
+                  child: StreamChatCore(
+                    client: client,
+                    onBackgroundEventReceived: widget.onBackgroundEventReceived,
+                    backgroundKeepAlive: widget.backgroundKeepAlive,
+                    connectivityStream: widget.connectivityStream,
+                    child: Builder(
+                      builder: (context) {
+                        StreamChatClient.additionalHeaders = {
+                          'X-Stream-Client':
+                              '${StreamChatClient.defaultUserAgent}-'
+                                  'ui-${StreamChatClient.packageVersion}',
+                        };
+                        return widget.child ?? const Offstage();
+                      },
+                    ),
                   ),
-                ),
-              );
-            },
+                );
+              },
+            ),
           ),
         ),
       ),
