@@ -4,6 +4,8 @@ import 'package:golden_toolkit/golden_toolkit.dart';
 import 'package:stream_chat_flutter/src/dialogs/confirmation_dialog.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+import '../material_app_wrapper.dart';
+
 void main() {
   testWidgets('ChannelInfoDialog shows info and members', (tester) async {
     await tester.pumpWidget(
@@ -38,7 +40,7 @@ void main() {
 
   testGoldens('golden test for ConfirmationDialog', (tester) async {
     await tester.pumpWidget(
-      MaterialApp(
+      MaterialAppWrapper(
         home: Scaffold(
           body: Builder(
             builder: (context) {
