@@ -4,6 +4,8 @@ import 'package:golden_toolkit/golden_toolkit.dart';
 import 'package:stream_chat_flutter/src/dialogs/message_dialog.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+import '../material_app_wrapper.dart';
+
 void main() {
   testWidgets('MessageDialog shows default info', (tester) async {
     await tester.pumpWidget(
@@ -57,7 +59,7 @@ void main() {
 
   testGoldens('golden test for default MessageDialog', (tester) async {
     await tester.pumpWidget(
-      MaterialApp(
+      MaterialAppWrapper(
         home: Scaffold(
           body: Builder(
             builder: (context) {
@@ -78,7 +80,7 @@ void main() {
 
   testGoldens('golden test for custom MessageDialog', (tester) async {
     await tester.pumpWidget(
-      MaterialApp(
+      MaterialAppWrapper(
         home: Scaffold(
           body: Builder(
             builder: (context) {
@@ -103,7 +105,7 @@ void main() {
   testGoldens('golden test for custom MessageDialog with no body',
       (tester) async {
     await tester.pumpWidget(
-      MaterialApp(
+      MaterialAppWrapper(
         home: Scaffold(
           body: Builder(
             builder: (context) {
