@@ -3,6 +3,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
 import 'package:stream_chat_flutter/src/message_input/command_button.dart';
 
+import '../material_app_wrapper.dart';
+
 void main() {
   testWidgets('CommandButton onPressed works', (tester) async {
     var count = 0;
@@ -27,9 +29,52 @@ void main() {
     expect(count, 1);
   });
 
-  testGoldens('golden test for CommandButton', (tester) async {
+  testWidgets('CommandButton should accept icon', (tester) async {
+    const icon = Icon(Icons.add);
     await tester.pumpWidget(
       MaterialApp(
+        home: Scaffold(
+          body: Center(
+            child: CommandButton(
+              icon: icon,
+              onPressed: () {},
+            ),
+          ),
+        ),
+      ),
+    );
+
+    final iconFound = find.byWidget(icon);
+    expect(iconFound, findsOneWidget);
+  });
+
+  testWidgets('CommandButton should not accept both color and icon',
+      (tester) async {
+    expect(
+      () => MaterialApp(
+        home: Scaffold(
+          body: Center(
+            child: CommandButton(
+              color: Colors.red,
+              icon: const Icon(Icons.add),
+              onPressed: () {},
+            ),
+          ),
+        ),
+      ),
+      throwsA(
+        isA<AssertionError>().having(
+          (e) => e.message,
+          'message',
+          'Either icon or color should be provided',
+        ),
+      ),
+    );
+  });
+
+  testGoldens('golden test for CommandButton', (tester) async {
+    await tester.pumpWidget(
+      MaterialAppWrapper(
         home: Scaffold(
           body: Center(
             child: CommandButton(
