@@ -3,6 +3,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+import '../material_app_wrapper.dart';
+
 void main() {
   testWidgets('CountdownButton works', (tester) async {
     await tester.pumpWidget(
@@ -23,7 +25,7 @@ void main() {
 
   testGoldens('golden test for CountdownButton', (tester) async {
     await tester.pumpWidget(
-      MaterialApp(
+      MaterialAppWrapper(
         home: StreamChatTheme(
           data: StreamChatThemeData.light(),
           child: const Scaffold(
