@@ -5,6 +5,7 @@ import 'package:mocktail/mocktail.dart';
 import 'package:stream_chat_flutter/src/message_widget/reactions/reaction_bubble.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+import '../material_app_wrapper.dart';
 import '../mocks.dart';
 
 void main() {
@@ -13,7 +14,9 @@ void main() {
     (WidgetTester tester) async {
       final client = MockClient();
       final clientState = MockClientState();
-      final themeData = ThemeData.light();
+      final themeData = ThemeData.light(
+        useMaterial3: false,
+      );
 
       when(() => client.state).thenReturn(clientState);
       when(() => clientState.currentUser).thenReturn(OwnUser(id: 'user-id'));
@@ -24,21 +27,28 @@ void main() {
           client: client,
           streamChatThemeData: theme,
           connectivityStream: Stream.value(ConnectivityResult.mobile),
-          child: SizedBox(
-            child: StreamReactionBubble(
-              reactions: [
-                Reaction(
-                  type: 'like',
-                  user: User(id: 'test'),
-                ),
-              ],
-              borderColor: theme.ownMessageTheme.reactionsBorderColor!,
-              backgroundColor: theme.ownMessageTheme.reactionsBackgroundColor!,
-              maskColor: theme.ownMessageTheme.reactionsMaskColor!,
+          child: Scaffold(
+            body: Center(
+              child: StreamReactionBubble(
+                reactions: [
+                  Reaction(
+                    type: 'like',
+                    user: User(id: 'test'),
+                  ),
+                ],
+                borderColor: theme.ownMessageTheme.reactionsBorderColor!,
+                backgroundColor:
+                    theme.ownMessageTheme.reactionsBackgroundColor!,
+                maskColor: theme.ownMessageTheme.reactionsMaskColor!,
+              ),
             ),
           ),
         ),
         surfaceSize: const Size(100, 100),
+        wrapper: (child) => MaterialAppWrapper(
+          theme: themeData,
+          home: child,
+        ),
       );
       await screenMatchesGolden(tester, 'reaction_bubble_like_light');
     },
@@ -176,7 +186,9 @@ void main() {
     (WidgetTester tester) async {
       final client = MockClient();
       final clientState = MockClientState();
-      final themeData = ThemeData();
+      final themeData = ThemeData(
+        useMaterial3: false,
+      );
 
       when(() => client.state).thenReturn(clientState);
       when(() => clientState.currentUser).thenReturn(OwnUser(id: 'user-id'));
@@ -186,32 +198,38 @@ void main() {
           client: client,
           connectivityStream: Stream.value(ConnectivityResult.mobile),
           streamChatThemeData: StreamChatThemeData.fromTheme(themeData),
-          child: SizedBox(
-            child: StreamReactionBubble(
-              reactions: [
-                Reaction(
-                  type: 'like',
-                  user: User(id: 'test'),
-                ),
-                Reaction(
-                  type: 'love',
-                  user: User(id: 'user-id'),
-                ),
-                Reaction(
-                  type: 'unknown',
-                  user: User(id: 'test'),
-                ),
-              ],
-              borderColor: Colors.red,
-              backgroundColor: Colors.blue,
-              maskColor: Colors.green,
-              reverse: true,
-              flipTail: true,
-              tailCirclesSpacing: 4,
+          child: Scaffold(
+            body: Center(
+              child: StreamReactionBubble(
+                reactions: [
+                  Reaction(
+                    type: 'like',
+                    user: User(id: 'test'),
+                  ),
+                  Reaction(
+                    type: 'love',
+                    user: User(id: 'user-id'),
+                  ),
+                  Reaction(
+                    type: 'unknown',
+                    user: User(id: 'test'),
+                  ),
+                ],
+                borderColor: Colors.red,
+                backgroundColor: Colors.blue,
+                maskColor: Colors.green,
+                reverse: true,
+                flipTail: true,
+                tailCirclesSpacing: 4,
+              ),
             ),
           ),
         ),
         surfaceSize: const Size(200, 200),
+        wrapper: (child) => MaterialAppWrapper(
+          theme: themeData,
+          home: child,
+        ),
       );
 
       await screenMatchesGolden(tester, 'reaction_bubble_2');
